+++ cli/src/commands/new.rs
@@ -132,7 +132,7 @@ pub(crate) struct NewArgs {
     /// Example: `jj new --insert-after A --insert-before D`:
     ///
     /// ```text
-    /// 
+    ///
     ///     D            D
     ///     |           / \
     ///     C          |   C